manually put my daily log stuff into the Notes.jsx and updated App.js to combine routes
